tasks/project4_genai/requirements.txt
pandas
python-dotenv
langchain-openai
langchain-core
openai
streamlit
jupyter
notebook
